reorganized /models
