+++ lib/illiad_client.rb
