+++ lib/publisher.rb
