+++ lib/utility.rb
