+++ models/fines/receipt.rb
