+++ spec/lib/authenticator_spec.rb
