+++ spec/lib/navigation_spec.rb
