+++ spec/lib/publisher_spec.rb
