+++ spec/lib/utility_spec.rb
