+++ spec/models/fines/fine_payer_spec.rb
