+++ spec/models/fines/receipt_spec.rb
